feat: add login process with JWT generation - add LoginController - add LoginDTO - add login endpoitn - add JWTResponseDTO - add security configuration class

+++ src/main/java/altamirano/hernandez/proyectogastos_springboot_angular/repositories/IUsuarioRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.Optional;
 
-public interface IUsuarioRepository extends CrudRepository<Usuario, Long> {
+public interface IUsuarioRepository extends CrudRepository<Usuario, Integer> {
 
     @Query("SELECT u FROM Usuario u WHERE u.email =:email")
     Optional<Usuario> findByEmail(@Param("email") String email);
