+++ src/main/java/altamirano/hernandez/proyectogastos_springboot_angular/services/interfaces/IUsuarioService.java
@@ -8,9 +8,9 @@ import java.util.Optional;
 
 public interface IUsuarioService {
     public abstract List<Usuario> findAll();
-    public abstract Usuario findById(Long id);
+    public abstract Usuario findById(int id);
     public Optional<Usuario> findByEmail(String email);
     public Optional<Usuario> findByEmailAndEstadoId(String email, Estado estado);
     public abstract void save(Usuario usuario);
-    public abstract void deleteById(Long id);
+    public abstract void deleteById(int id);
 }
